Changed Tabs to Space, moved classes, and edited lang file
